+++ server/console_leaderboard.go
@@ -142,7 +142,7 @@ func (s *ConsoleServer) DeleteLeaderboard(ctx context.Context, in *console.Leade
 		return nil, status.Error(codes.InvalidArgument, "Expects a leaderboard ID")
 	}
 
-	if err := s.leaderboardCache.Delete(ctx, in.Id); err != nil {
+	if err := s.leaderboardCache.Delete(ctx, s.leaderboardRankCache, s.leaderboardScheduler, in.Id); err != nil {
 		// Logged internally
 		return nil, status.Error(codes.Internal, "Error deleting leaderboard.")
 	}